Revised spa-with-api tests to include failing run_worker_first tests
