+++ packages/vite-plugin-cloudflare/playground/spa-with-api/__tests__/https/spa-with-api.spec.ts
@@ -1,7 +1,1 @@
-import { expect, test } from "vitest";
-import { page } from "../../../__test-utils__";
-
-test("returns the correct home page", async () => {
-	const content = await page.textContent("h1");
-	expect(content).toBe("Vite + React");
-});
+import "../base-tests";
