+++ packages/vite-plugin-cloudflare/playground/spa-with-api/api/some-file.txt
@@ -1,0 +1,1 @@
+File content.
