+++ packages/vite-plugin-cloudflare/playground/spa-with-api/wrangler.run-worker-first.jsonc
